+++ src/storage/index.ts
@@ -69,7 +69,6 @@ const defaultSettings = {
   novelTtsEnabled: false,
   novelTtsVoice: "M5" as "M5" | "F5",
   novelTtsSteps: 5,
-  novelTtsThreads: "4",
   storyCardAutocompleteTypes: MENTION_CARD_TYPES.map(type => type.value),
 
   trinetraApiKey: "",